add chars to clan

+++ src/controllers/character.js
@@ -65,29 +65,23 @@ import { GalleryCharacters } from '../db/models/gallery.js';
 import { CharactersCollection } from '../db/models/character.js';
 
 export async function deleteCharacterController(req, res) {
-  // 🔹 Спочатку шукаємо персонажа
   const character = await CharactersCollection.findOne({
     _id: req.params.charId,
     userId: req.user.id,
   });
 
-  // 🔹 Якщо персонаж не знайдений
   if (!character) {
     throw new createHttpError.NotFound('Character not found');
   }
 
-  // 🔹 Якщо персонаж — лідер клану, не дозволяємо видалити
   if (character.clan?.role === 'leader') {
-    return res.status(403).json({
-      message:
-        'Лідер не може видалити персонажа, поки не передасть керівництво кланом',
-    });
+    return;
   }
-
-  // 🔹 Видаляємо персонажа
+  if (character.clan?.clanId) {
+    return;
+  }
   await deleteCharacter(req.params.charId, req.user.id);
 
-  // 🔹 Видаляємо копії з галереї
   await GalleryCharacters.deleteMany({ originalCharId: req.params.charId });
 
   res.status(204).end();
@@ -113,12 +107,7 @@ export async function getAllCarsController(req, res) {
 export const getFreeCharsController = async (req, res) => {
   const userId = req.user.id;
   const clanServer = req.query.server;
-  console.log('🚀 ~ getFreeCharsController ~ body:', req.query);
-  const charsAll = await CharactersCollection.find({
-    userId,
-    server: clanServer,
-  });
-  console.log('All chars on this server:', charsAll);
+
   const chars = await CharactersCollection.find({
     userId,
     server: clanServer,
